Refactor: split beamer locations into separate files
