+++ lib/beamer/beamer_app.dart
@@ -4,8 +4,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:form_builder_validators/localization/l10n.dart';
-import 'package:lotti/beamer/beamer_locations.dart';
-import 'package:lotti/beamer/locations/settings_locations.dart';
+import 'package:lotti/beamer/locations/dashboards_location.dart';
+import 'package:lotti/beamer/locations/journal_location.dart';
+import 'package:lotti/beamer/locations/settings_location.dart';
+import 'package:lotti/beamer/locations/tasks_location.dart';
 import 'package:lotti/blocs/audio/player_cubit.dart';
 import 'package:lotti/blocs/audio/recorder_cubit.dart';
 import 'package:lotti/blocs/sync/outbox_cubit.dart';
